# backend/docs/swagger.yaml
basePath: /
definitions:
...
  gin.H:
    additionalProperties: {}
    type: object
  models.Application:
    properties:
      assetCriticality:
        type: string
      createdAt:
        type: string
      description:
        type: string
      groupId:
        type: integer
      id:
        type: integer
      name:
        type: string
      repositoryUrl:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
      versions:
        items:
          $ref: '#/definitions/models.ApplicationVersion'
        type: array
    type: object
  models.ApplicationVersion:
    properties:
      applicationId:
        type: integer
      branch:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      name:
        type: string
      tag:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@servasec.local
    name: API Support
  description: Security scanner management platform API
  title: Servasec API
  version: "1.0"
paths:
...
  /applications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of applications
          schema:
            items:
              $ref: '#/definitions/models.Application'
            type: array
      summary: List applications
      tags:
      - Applications
    post:
      consumes:
      - application/json
      parameters:
      - description: Application details
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created application
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create application
      tags:
      - Applications
  /applications/{id}:
    delete:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Application deleted
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete application
      tags:
      - Applications
    get:
      parameters:
      - description: Application ID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Application with default version
          schema:
            type: object
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get application
      tags:
      - Applications
    put:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated application
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update application
      tags:
      - Applications
...
  /applications/{id}/regenerate-token:
    post:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: New API token
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Regenerate API token
      tags:
      - Applications
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"